tint/writer/hlsl: Inline constant expressions
This is required to handle materialized values, and for constant
expressions.

Bug: tint:1504
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/92081
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Ben Clayton <bclayton@chromium.org>

+++ test/tint/array/assign_to_storage_var.wgsl.expected.hlsl
@@ -88,7 +88,7 @@ void tint_symbol_9(RWByteAddressBuffer buffer, uint offset, int value[4][3][2]) 
 
 void foo(int4 src_param[4]) {
   int4 src_function[4] = (int4[4])0;
-  const int4 tint_symbol_15[4] = {int4((1).xxxx), int4((2).xxxx), int4((3).xxxx), int4((3).xxxx)};
+  const int4 tint_symbol_15[4] = {(1).xxxx, (2).xxxx, (3).xxxx, (3).xxxx};
   tint_symbol_3(tint_symbol, 0u, tint_symbol_15);
   tint_symbol_3(tint_symbol, 0u, src_param);
   const int4 tint_symbol_1[4] = ret_arr();
