+++ test/tint/bug/tint/1385.wgsl.expected.hlsl
@@ -1,7 +1,7 @@
 ByteAddressBuffer data : register(t1, space0);
 
 int foo() {
-  return asint(data.Load((4u * uint(0))));
+  return asint(data.Load((4u * 0u)));
 }
 
 [numthreads(16, 16, 1)]
